Sun, Feb 12, 2023, 5:16 PM +01:00

+++ taqueria/taqueria.py
@@ -15,7 +15,7 @@ def main():
     while True:
         try:
             item = input("Item: ")
-            total = int(taq[item])
+            total = float(taq[item])
             print(f"Total: ${total}")
         except EOFError:
             print("")
